clean up a bit

+++ examples/demo-shell/demo_compositor.h
@@ -51,8 +51,8 @@ private:
     graphics::DisplayBuffer& display_buffer;
     std::shared_ptr<compositor::Scene> const scene;
     std::shared_ptr<compositor::CompositorReport> const report;
-    float const shadow_radius;
-    float const titlebar_height;
+    int const shadow_radius;
+    int const titlebar_height;
     DemoRenderer renderer;
 };
